src/MCP/energy/requirements.txt
httpx
uvicorn
mcp
python-dotenv==1.0.0
requests
fastmcp
